pyproject.toml
[tool.poetry]
name = "absa-nlp"
version = "0.1.0"
description = ""
authors = ["Quan <toolowner@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.7.0"
openai = "^1.17.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
